requirements.txt
openai
tiktoken
langchain
langchain-openai
chromadb
sentence_transformers
unstructured
pypdf
pinecone-client
python-dotenv
faiss-cpu
langchain_experimental
tabulate
nest-asyncio
